+++ app/src/main/java/ru/practicum/android/diploma/ui/screens/searchfragment/SearchFragment.kt
@@ -155,4 +155,3 @@ fun SearchScreenServerErrorPreview() {
         )
     }
 }
-
